Final warning resolutions on Linux

+++ xash3d_engine/engine/client/cl_netgraph.c
@@ -282,7 +282,7 @@ static void NetGraph_DrawTimes(wrect_t rect, int x, int w)
 
 			for ( j = start; j < h; j++ )
 			{
-				int color = NETGRAPH_NET_COLORS + j + extrap_point;
+				size_t color = (size_t)(NETGRAPH_NET_COLORS + j + extrap_point);
 				color = Q_min(color, ARRAYSIZE(netcolors) - 1);
 
 				NetGraph_DrawRect(&fill, netcolors[color]);
@@ -301,7 +301,7 @@ static void NetGraph_DrawTimes(wrect_t rect, int x, int w)
 
 			for ( j = 0; j < h; j++ )
 			{
-				int color = NETGRAPH_NET_COLORS + j + oldh;
+				size_t color = (size_t)(NETGRAPH_NET_COLORS + j + oldh);
 				color = Q_min(color, ARRAYSIZE(netcolors) - 1);
 
 				NetGraph_DrawRect(&fill, netcolors[color]);
